Convert to TypeScript
This involved a lot of pulling of hair and gnashing of teeth. It turns
out that trying to represent the filters and filter selections as types
and have them coordinated between each other involves some fancy
TypeScript tricks. Notably, the filter types cannot be merged
successfully, so discriminated union types need to be utilized in order
to write completely type-safe code.